# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Human Dashboard API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/qb-session-login-status/{session_id}": {
            "get": {
                "tags": [
                    "QB Logout"
                ],
                "summary": "Check QB login status",
                "description": "Returns **200** when a QB user is bound to the given mobile session UUID. Returns **401** when no QB has this `session_id` (Postman **Check QB Login Status 200** / invalid session **Check QB Login Status 403** — API uses **401 Unauthenticated**).",
                "operationId": "qbSessionLoginStatus",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "description": "Mobile pairing session UUID from `POST /api/mobile/create-session`",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "822fc835-75aa-48bf-8473-354a4913aab2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QB is linked to this session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "session_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "logged_in": {
                                            "description": "Mirrors the QB user's `is_loggin` flag in the database.",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated — no QB user with this session_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum bearer token. Use format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}